Rename utils/helpers.js file to utils/stateChecks.js
