+++ src/utils/stateChecks.js
